pyproject.toml
[tool.poetry]
name = "adeo-tech-test-httpx-async"
version = "0.1.0"
description = ""
authors = ["yohanv <support@vachet.me>"]
readme = "README.md"
packages = [{include = "test_httpx_async"}]

[tool.poetry.dependencies]
python = "^3.13"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
flake8 = "7.1.1"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
